rearrangle `hydroflow_plus_test` examples into individual files
