+++ hydroflow_plus_test/src/bin/compute_pi.rs
@@ -3,9 +3,8 @@
 async fn main() {
     let batch_size = 8192;
 
-    hydroflow_plus::launch!(|ports| hydroflow_plus_test::cluster::compute_pi_runtime!(
-        ports,
-        &batch_size,
-    ))
+    hydroflow_plus::launch!(
+        |ports| hydroflow_plus_test::cluster::compute_pi::compute_pi_runtime!(ports, &batch_size,)
+    )
     .await;
 }
