+++ hydroflow_plus_test/src/bin/many_to_many.rs
@@ -2,7 +2,7 @@
 #[tokio::main]
 async fn main() {
     hydroflow_plus::util::cli::launch!(|ports| {
-        hydroflow_plus_test::cluster::many_to_many_runtime!(ports)
+        hydroflow_plus_test::cluster::many_to_many::many_to_many_runtime!(ports)
     })
     .await;
 }
