+++ hydroflow_plus_test/src/bin/map_reduce.rs
@@ -2,7 +2,7 @@
 #[tokio::main]
 async fn main() {
     hydroflow_plus::util::cli::launch!(|ports| {
-        hydroflow_plus_test::cluster::map_reduce_runtime!(ports)
+        hydroflow_plus_test::cluster::map_reduce::map_reduce_runtime!(ports)
     })
     .await;
 }
