+++ hydroflow_plus_test/src/bin/simple_cluster.rs
@@ -2,7 +2,7 @@
 #[tokio::main]
 async fn main() {
     hydroflow_plus::util::cli::launch!(|ports| {
-        hydroflow_plus_test::cluster::simple_cluster_runtime!(ports)
+        hydroflow_plus_test::cluster::simple_cluster::simple_cluster_runtime!(ports)
     })
     .await;
 }
