+++ hydroflow_plus_test/src/distributed/first_ten.rs
@@ -2,23 +2,6 @@ use hydroflow_plus::*;
 use serde::{Deserialize, Serialize};
 use stageleft::*;
 
-pub fn first_ten<'a, D: LocalDeploy<'a>>(
-    flow: &FlowBuilder<'a, D>,
-    process_spec: &impl ProcessSpec<'a, D>,
-) {
-    let process = flow.process(process_spec);
-    let numbers = flow.source_iter(&process, q!(0..10));
-    numbers.for_each(q!(|n| println!("{}", n)));
-}
-
-#[stageleft::entry]
-pub fn first_ten_runtime<'a>(
-    flow: FlowBuilder<'a, SingleProcessGraph>,
-) -> impl Quoted<'a, Hydroflow<'a>> {
-    first_ten(&flow, &());
-    flow.extract().optimize_default()
-}
-
 #[derive(Serialize, Deserialize)]
 struct SendOverNetwork {
     pub n: u32,
@@ -59,11 +42,6 @@ pub fn first_ten_distributed_runtime<'a>(
 mod tests {
     use hydro_deploy::{Deployment, HydroflowCrate};
     use hydroflow_plus_cli_integration::{DeployCrateWrapper, DeployProcessSpec};
-
-    #[test]
-    fn instantiate_first_ten() {
-        let _ = super::first_ten_runtime!();
-    }
 
     #[tokio::test]
     async fn first_ten_distributed() {
