+++ hydroflow_plus_test/src/local/mod.rs
@@ -1,0 +1,4 @@
+pub mod compute_pi;
+pub mod first_ten;
+pub mod futures;
+pub mod negation;
